Fix: centralize normalization and diff logic in reusable function
